refactor: delete redundant cast

+++ engine/src/vk_pipelines.cpp
@@ -10,7 +10,7 @@ bool load_shader_module(const char *filePath, VkDevice device, VkShaderModule *o
         return false;
     }
 
-    const size_t fileSize = (size_t)file.tellg();
+    const size_t fileSize = file.tellg();
 
     std::vector<uint32_t> buffer(fileSize / sizeof(uint32_t));
